requirements.txt
fastapi
pydantic[email]
uvicorn[standard]
python-dotenv
slowapi